# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API CODEPLEX",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión válido"
                    },
                    "400": {
                        "$ref": "#/components/responses/Authentication"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtener información de usuario autenticado",
                "operationId": "authMe",
                "responses": {
                    "200": {
                        "description": "Información de usuario obtenida"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar usuario",
                "operationId": "authRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Información de usuario obtenida"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/auth/validate-token/{token}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Validar token",
                "operationId": "authValidateToken",
                "parameters": [
                    {
                        "parameter": "token",
                        "name": "token",
                        "in": "path",
                        "description": "Token de autenticación",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token validado"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear Producto",
                "operationId": "productStore",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "400": {
                        "$ref": "#/components/responses/Authentication"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/api/v1/products/{id}?_method=PUT": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar producto",
                "operationId": "productUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de producto",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto actualizado"
                    },
                    "400": {
                        "$ref": "#/components/responses/Authentication"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "JsonApiDefinition": {
                "title": "JsonApiDefinition",
                "description": "JsonApiDefinition",
                "properties": {
                    "environment": {
                        "type": "string",
                        "example": "local"
                    },
                    "version": {
                        "type": "string",
                        "example": "1.0.0"
                    },
                    "name": {
                        "type": "string",
                        "example": "WPARK Api"
                    },
                    "summary": {
                        "type": "string",
                        "example": "Api for obtain information on operations within a Compound."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JsonApiDefinition"
                }
            },
            "ValidationErrorsDefinition": {
                "title": "ValidationErrorsDefinition",
                "description": "ValidationErrorsDefinition",
                "properties": {
                    "field": {
                        "type": "string",
                        "example": ""
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ValidationErrorsDefinition"
                }
            },
            "LoginRequest": {
                "title": "Login Request",
                "description": "Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email de usuario",
                        "type": "string",
                        "maxLength": 255,
                        "example": "giluchi8@gmail.com"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "maxLength": 100,
                        "example": "secret_password"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LoginRequest"
                }
            },
            "RegisterRequest": {
                "title": "Register Request",
                "description": "Register request body data",
                "required": [
                    "idplan",
                    "nombre",
                    "email",
                    "password",
                    "ruc",
                    "razon",
                    "telefono"
                ],
                "properties": {
                    "idplan": {
                        "description": "Id del plan contratado por la empresa",
                        "type": "integer",
                        "example": "1"
                    },
                    "nombre": {
                        "description": "Nombre de usuario",
                        "type": "string",
                        "maxLength": 100,
                        "example": "donpepito"
                    },
                    "email": {
                        "description": "Email de la empresa",
                        "type": "string",
                        "maxLength": 50,
                        "example": "empresa@gmail.com"
                    },
                    "password": {
                        "description": "Contraseña de acceso al sistema para la empresa",
                        "type": "string",
                        "minLength": 6,
                        "example": "secret"
                    },
                    "ruc": {
                        "description": "RUC de la empresa",
                        "type": "string",
                        "maxLength": 11,
                        "minLength": 11,
                        "example": "10201239871"
                    },
                    "razon": {
                        "description": "Razón social de la empresa",
                        "type": "string",
                        "maxLength": 100,
                        "example": "Don Pepito SA"
                    },
                    "telefono": {
                        "description": "Teléfono de la empresa",
                        "type": "string",
                        "maxLength": 100,
                        "example": "963896712"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RegisterRequest"
                }
            },
            "ProductStoreRequest": {
                "title": "Product Store Request",
                "description": "Product store request form data",
                "required": [
                    "idempresa",
                    "idlinea",
                    "idlineasub",
                    "idtipoproducto",
                    "codigo",
                    "descripcion"
                ],
                "properties": {
                    "idempresa": {
                        "description": "Id de la empresa",
                        "type": "string",
                        "example": "1-20539782232"
                    },
                    "idlinea": {
                        "description": "Id de la línea",
                        "type": "integer",
                        "example": "1"
                    },
                    "idlineasub": {
                        "description": "Id de la sub-línea",
                        "type": "integer",
                        "example": "1"
                    },
                    "idtipoproducto": {
                        "description": "Id del tipo de producto",
                        "type": "integer",
                        "example": "1"
                    },
                    "idsunatt07": {
                        "description": "Id Sunat 07",
                        "type": "string",
                        "maxLength": 3,
                        "example": ""
                    },
                    "codigo": {
                        "description": "Código de producto",
                        "type": "string",
                        "maxLength": 50,
                        "example": "2257"
                    },
                    "descripcion": {
                        "description": "Nombre o descripción del producto",
                        "type": "string",
                        "maxLength": 3000,
                        "example": "Chocolates"
                    },
                    "activo": {
                        "description": "¿El producto está activo?. Activo: 1, Inactivo: 0",
                        "type": "string",
                        "default": "1",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "infad1": {
                        "description": "",
                        "type": "string",
                        "maxLength": 3000,
                        "example": ""
                    },
                    "infad2": {
                        "description": "",
                        "type": "string",
                        "maxLength": 3000,
                        "example": ""
                    },
                    "infad3": {
                        "description": "",
                        "type": "string",
                        "maxLength": 3000,
                        "example": ""
                    },
                    "porpercepcion": {
                        "description": "",
                        "type": "integer",
                        "example": ""
                    },
                    "porisc": {
                        "description": "",
                        "type": "integer",
                        "example": ""
                    },
                    "estadoventa": {
                        "description": "Estado de venta. Activo: 1, Inactivo: 0",
                        "type": "string",
                        "default": "1",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "escombo": {
                        "description": "SI: 1, NO: 0",
                        "type": "string",
                        "default": "0",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "icbper": {
                        "description": "SI: 1, NO: 0",
                        "type": "string",
                        "default": "0",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "urlapi": {
                        "description": "",
                        "type": "string",
                        "example": ""
                    },
                    "itemPrecios": {
                        "description": "Listado de precios. Debe ser un JSON convertido a string",
                        "type": "string",
                        "example": ""
                    },
                    "upload_path": {
                        "description": "Ruta donde se almacenan las imágenes del producto",
                        "type": "string",
                        "example": ""
                    },
                    "imagen1": {
                        "description": "Imagen nro 1",
                        "type": "file",
                        "example": ""
                    },
                    "imagen2": {
                        "description": "Imagen nro 2",
                        "type": "file",
                        "example": ""
                    },
                    "imagen3": {
                        "description": "Imagen nro 3",
                        "type": "file",
                        "example": ""
                    },
                    "imagen4": {
                        "description": "Imagen nro 4",
                        "type": "file",
                        "example": ""
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductStoreRequest"
                }
            },
            "ProductUpdateRequest": {
                "title": "Product Update Request",
                "description": "Product update request form data",
                "required": [
                    "idempresa",
                    "idlinea",
                    "idlineasub",
                    "idtipoproducto",
                    "codigo",
                    "descripcion"
                ],
                "properties": {
                    "idempresa": {
                        "description": "Id de la empresa",
                        "type": "string",
                        "example": "1-20539782232"
                    },
                    "idlinea": {
                        "description": "Id de la línea",
                        "type": "integer",
                        "example": "1"
                    },
                    "idlineasub": {
                        "description": "Id de la sub-línea",
                        "type": "integer",
                        "example": "1"
                    },
                    "idtipoproducto": {
                        "description": "Id del tipo de producto",
                        "type": "integer",
                        "example": "1"
                    },
                    "idsunatt07": {
                        "description": "Id Sunat 07",
                        "type": "string",
                        "maxLength": 3,
                        "example": ""
                    },
                    "codigo": {
                        "description": "Código de producto",
                        "type": "string",
                        "maxLength": 50,
                        "example": "2257"
                    },
                    "descripcion": {
                        "description": "Nombre o descripción del producto",
                        "type": "string",
                        "maxLength": 3000,
                        "example": "Chocolates"
                    },
                    "activo": {
                        "description": "¿El producto está activo?. Activo: 1, Inactivo: 0",
                        "type": "string",
                        "default": "1",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "infad1": {
                        "description": "",
                        "type": "string",
                        "maxLength": 3000,
                        "example": ""
                    },
                    "infad2": {
                        "description": "",
                        "type": "string",
                        "maxLength": 3000,
                        "example": ""
                    },
                    "infad3": {
                        "description": "",
                        "type": "string",
                        "maxLength": 3000,
                        "example": ""
                    },
                    "porpercepcion": {
                        "description": "",
                        "type": "integer",
                        "example": ""
                    },
                    "porisc": {
                        "description": "",
                        "type": "integer",
                        "example": ""
                    },
                    "estadoventa": {
                        "description": "Estado de venta. Activo: 1, Inactivo: 0",
                        "type": "string",
                        "default": "1",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "escombo": {
                        "description": "SI: 1, NO: 0",
                        "type": "string",
                        "default": "0",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "icbper": {
                        "description": "SI: 1, NO: 0",
                        "type": "string",
                        "default": "0",
                        "enum": [
                            "1",
                            "0"
                        ]
                    },
                    "urlapi": {
                        "description": "",
                        "type": "string",
                        "example": ""
                    },
                    "itemPrecios": {
                        "description": "Listado de precios. Debe ser un JSON convertido a string",
                        "type": "string",
                        "example": ""
                    },
                    "upload_path": {
                        "description": "Ruta donde se almacenan las imágenes del producto",
                        "type": "string",
                        "example": ""
                    },
                    "imagen1": {
                        "description": "Imagen nro 1",
                        "type": "file",
                        "example": ""
                    },
                    "imagen2": {
                        "description": "Imagen nro 2",
                        "type": "file",
                        "example": ""
                    },
                    "imagen3": {
                        "description": "Imagen nro 3",
                        "type": "file",
                        "example": ""
                    },
                    "imagen4": {
                        "description": "Imagen nro 4",
                        "type": "file",
                        "example": ""
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductUpdateRequest"
                }
            },
            "UnauthorizedJsonContent": {
                "title": "Unauthorized Json",
                "description": "Unauthorized json response content",
                "properties": {
                    "jsonapi": {
                        "$ref": "#/components/schemas/JsonApiDefinition"
                    },
                    "code": {
                        "type": "integer",
                        "example": "401"
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UnauthorizedJson"
                }
            }
        },
        "responses": {
            "Authentication": {
                "description": "Authentication Error."
            },
            "BadRequest": {
                "description": "Bad Request Error."
            },
            "Forbidden": {
                "description": "Permission Error."
            },
            "InternalServerError": {
                "description": "Internal Server Error"
            },
            "NotFound": {
                "description": "Resource Not Found."
            },
            "Unauthorized": {
                "description": "Unauthorized Error."
            },
            "UnprocessableEntity": {
                "description": "Validation Error."
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "id, eg; 1",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}